requirements.txt
boto3==1.40.4
joblib==1.5.2
lightgbm==3.3.5
numpy==1.26.4
pandas==2.2.3
requests==2.32.3
pymysql
psycopg2-binary==2.9.9